+++ .pre-commit-config.yaml
@@ -6,7 +6,9 @@ exclude: >
       \.vscode/settings\.json|
       docs/source/conf\.py|
       tests/test_exercise/.*|
+      tests/test_exercise_references/.*|
       tests/test_solution/.*|
+      tests/test_solution_references/.*|
       tests/test_hiddendirective/.*|
     )$
