+++ tests/test_exercise_references/_enum_numref_placeholders.xml
