+++ tests/test_solution_references.py
@@ -6,43 +6,43 @@ import pytest
 @pytest.mark.parametrize(
     "idir",
     [
-        "_linked_enum.html",
-        "_linked_enum_class.html",
-        "_linked_unenum_mathtitle.html",
-        "_linked_unenum_mathtitle2.html",
-        "_linked_unenum_notitle.html",
-        "_linked_unenum_title.html",
-        "_linked_wrong_targetlabel.html",
+        "_linked_ref_enum.html",
+        "_linked_ref_unenum_mathtitle.html",
+        "_linked_ref_unenum_mathtitle2.html",
+        "_linked_ref_unenum_notitle.html",
+        "_linked_ref_unenum_title.html",
     ],
 )
-def test_solution(app, idir, file_regression):
-    """Test solution directive markup."""
-    app.build()
+def test_reference(app, idir, file_regression):
+    """Test solution ref role markup."""
+    app.builder.build_all()
     path_solution_directive = app.outdir / "solution" / idir
     assert path_solution_directive.exists()
-
     # get content markup
     soup = BeautifulSoup(
         path_solution_directive.read_text(encoding="utf8"), "html.parser"
     )
 
-    sol = soup.select("div.solution")[0]
-    file_regression.check(str(sol), basename=idir.split(".")[0], extension=".html")
+    excs = ""
+    all_refs = soup.select("p")
+    for ref in all_refs:
+        excs += f"{ref}\n"
+
+    file_regression.check(
+        str(excs[:-1]), basename=idir.split(".")[0], extension=".html"
+    )
 
 
 @pytest.mark.sphinx("html", testroot="mybook")
 @pytest.mark.parametrize(
     "docname",
     [
-        "_linked_duplicate_label",
-        "_linked_enum",
-        "_linked_enum_class",
-        "_linked_missing_arg",
-        "_linked_unenum_mathtitle",
-        "_linked_unenum_mathtitle2",
-        "_linked_unenum_notitle",
-        "_linked_unenum_title",
-        "_linked_wrong_targetlabel",
+        "_linked_ref_enum",
+        "_linked_ref_unenum_mathtitle",
+        "_linked_ref_unenum_mathtitle2",
+        "_linked_ref_unenum_notitle",
+        "_linked_ref_unenum_title",
+        "_linked_ref_wronglabel",
     ],
 )
 def test_solution_doctree(app, docname, file_regression, get_sphinx_app_doctree):
