+++ tests/test_solution_references/_linked_ref_unenum_mathtitle2.xml
