+++ tests/test_solution_references/_linked_ref_unenum_title.xml
